paddle ocr test added
