+++ chunkmydocs/src/utils/services/mod.rs
@@ -5,9 +5,9 @@ pub mod images;
 pub mod llm;
 pub mod log;
 pub mod ocr;
-pub mod ocr_services;
 pub mod payload;
 pub mod pdf;
 pub mod pdla;
 pub mod search;
+pub mod segment_ocr;
 pub mod structured_extraction;
