+++ compose.yaml
@@ -163,7 +163,7 @@ services:
     env_file:
       - .env
     deploy:
-      replicas: 1
+      replicas: 0
     restart: always
 
   structured-extraction:
